Rework goal reminder notifications
Removed original concept of morning, mid-day, and evening reminders. There are now three generic, configurable reminder options that can be enabled/disabled based on a user's preference. The limit can be easily changed in the future if need be.

The notifications will also overwrite the previous to reduce clutter.

Reminders also tell the user if they have incomplete goal reflections.